refactor(simulation): move simulation web dtos by feature
