+++ src/main/java/com/renewsim/backend/simulation_service/create/web/CreateSimulationController.java
@@ -3,9 +3,9 @@ package com.renewsim.backend.simulation_service.create.web;
 import com.renewsim.backend.simulation_service.create.application.CreateRealSimulationUseCase;
 import com.renewsim.backend.simulation_service.shared.web.SimulationRequestContext;
 import com.renewsim.backend.simulation_service.shared.web.SimulationRequestContextFactory;
+import com.renewsim.backend.simulation_service.detail.web.dto.SimulationDetailsResponseDTO;
 import com.renewsim.backend.simulation_service.web.controller.SimulationWebMapper;
-import com.renewsim.backend.simulation_service.web.dto.CreateSolarSimulationRequestDTO;
-import com.renewsim.backend.simulation_service.web.dto.SimulationDetailsResponseDTO;
+import com.renewsim.backend.simulation_service.create.web.dto.CreateSolarSimulationRequestDTO;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.tags.Tag;
 import jakarta.validation.Valid;
