+++ src/main/java/com/renewsim/backend/simulation_service/dashboard/web/SimulationDashboardController.java
@@ -3,8 +3,8 @@ package com.renewsim.backend.simulation_service.dashboard.web;
 import com.renewsim.backend.simulation_service.application.dashboard.GetPortfolioDashboardUseCase;
 import com.renewsim.backend.simulation_service.shared.web.SimulationRequestContext;
 import com.renewsim.backend.simulation_service.shared.web.SimulationRequestContextFactory;
+import com.renewsim.backend.simulation_service.dashboard.web.dto.PortfolioDashboardResponseDTO;
 import com.renewsim.backend.simulation_service.web.controller.SimulationWebMapper;
-import com.renewsim.backend.simulation_service.web.dto.PortfolioDashboardResponseDTO;
 import io.swagger.v3.oas.annotations.Operation;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.ResponseEntity;
