+++ src/main/java/com/renewsim/backend/simulation_service/dashboard/web/dto/PortfolioDashboardDistributionByTechnologyDTO.java
@@ -1,4 +1,4 @@
-package com.renewsim.backend.simulation_service.web.dto;
+package com.renewsim.backend.simulation_service.dashboard.web.dto;
 
 public record PortfolioDashboardDistributionByTechnologyDTO(
                 String label,
