+++ src/main/java/com/renewsim/backend/simulation_service/dashboard/web/dto/PortfolioDashboardPrioritizedScenarioDTO.java
@@ -1,19 +1,15 @@
-package com.renewsim.backend.simulation_service.web.dto;
+package com.renewsim.backend.simulation_service.dashboard.web.dto;
 
-import java.util.List;
-
-public record PortfolioDashboardRecommendedScenarioDTO(
+public record PortfolioDashboardPrioritizedScenarioDTO(
                 String id,
                 String name,
                 String technology,
+                String status,
                 String location,
                 Double roiPercent,
                 Double paybackYears,
                 Double capex,
                 Double estimatedAnnualSavings,
                 String priority,
-                String headline,
-                List<String> drivers,
-                String mainRisk,
-                String nextStep) {
+                Integer score) {
 }
