+++ src/test/java/com/renewsim/backend/simulation_service/detail/web/SimulationDetailControllerTest.java
@@ -3,7 +3,7 @@ package com.renewsim.backend.simulation_service.detail.web;
 import com.renewsim.backend.auth_service.domain.AuthenticatedUser;
 import com.renewsim.backend.simulation_service.application.detailSimulation.GetRealSimulationUseCase;
 import com.renewsim.backend.simulation_service.application.shared.SimulationDetailsResult;
-import com.renewsim.backend.simulation_service.web.dto.SimulationDetailsResponseDTO;
+import com.renewsim.backend.simulation_service.detail.web.dto.SimulationDetailsResponseDTO;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
